used eslint to refactor
