+++ front-end/src/components/Course.jsx
@@ -1,10 +1,10 @@
-import React from 'react'
+import React from 'react';
 
 const Course = ({ match }) => (
-    <div>
-      <h3>{match.params.code}</h3>
-      <h3>{match.url}</h3>
-    </div>
-  );
+  <div>
+    <h3>{match.params.code}</h3>
+    <h3>{match.url}</h3>
+  </div>
+);
 
-  export default Course;
+export default Course;
